+++ database/migrations/2023_02_15_194241_create_groups_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('groups', function (Blueprint $table) {
             $table->id();
 			$table->string('title');
-			$table->unsignedTinyInteger('crew_id');
+			$table->unsignedBigInteger('crew_id');
             $table->timestamps();
         });
     }
